Rename variable with confusing name
I saw this gem of not so legible code in solib-svr4.c (scan_dyntag):

	if (dyn_tag == dyntag)

and thought it deserved a small rename.

This just renames variables to be a bit more clear for those who read the
code. I also constified the parameter because, why not. The same was
done in scan_dyntag_auxv as well.

Tested only by rebuilding, since the change was done mechanically.

gdb/Changelog:

2014-08-01  Simon Marchi  <simon.marchi@ericsson.com>

	* solib-svr4.c (scan_dyntag): Rename dyntag and dyn_tag variables.
	(scan_dyntag_auxv): Same.